add resource update tests, force-new on sign, use OCI validators
