+++ .github/workflows/test.yml
@@ -14,7 +14,7 @@ jobs:
         ref: refs/pull/${{ github.event.pull_request.number }}/merge
     - uses: actions/setup-go@v4
       with:
-        go-version: '1.18'
+        go-version: 1.20.x
     - run: go generate ./...
     - name: git diff
       run: |
@@ -43,22 +43,11 @@ jobs:
         ref: refs/pull/${{ github.event.pull_request.number }}/merge
     - uses: actions/setup-go@v4
       with:
-        go-version: '1.18'
-
-    - uses: imjasonh/setup-crane@v0.3
-    - uses: chainguard-dev/actions/setup-registry@main
-      with:
-        port: 5000
-
+        go-version: 1.20.x
     - uses: hashicorp/setup-terraform@v2
       with:
         terraform_version: ${{ matrix.terraform }}
         terraform_wrapper: false
-
     - run: go mod download
     - run: go build -v .
-    - run: |
-        crane copy --insecure cgr.dev/chainguard/static:latest-glibc localhost:5000/cosign-testing
-        export TEST_IMAGE=localhost:5000/cosign-testing@$(crane digest --insecure localhost:5000/cosign-testing)
-
-        TF_ACC=1 go test -v -cover ./internal/provider/
+    - run: TF_ACC=1 go test -v -cover ./internal/provider/
